Change Integer to Long

+++ engine/components-api/src/main/java/com/cloud/ha/HighAvailabilityManager.java
@@ -32,17 +32,17 @@ import java.util.List;
  */
 public interface HighAvailabilityManager extends Manager {
 
-    ConfigKey<Integer> TimeBetweenCleanup = new ConfigKey<>("Advanced", Integer.class,
+    ConfigKey<Long> TimeBetweenCleanup = new ConfigKey<>("Advanced", Long.class,
         "time.between.cleanup", valueOf(3600 * 24), "Time in seconds to wait before the cleanup thread runs.",
         false, null);
 
     ConfigKey<Integer> MaxRetries = new ConfigKey<>("Advanced", Integer.class, "max.retries",
         valueOf(5), "Number of times to retry start.", false, null);
 
-    ConfigKey<Integer> TimeToSleep = new ConfigKey<>("Advanced", Integer.class, "time.to.sleep",
+    ConfigKey<Long> TimeToSleep = new ConfigKey<>("Advanced", Long.class, "time.to.sleep",
         valueOf(60), "Time in seconds to sleep if no work items are found.", false, null);
 
-    ConfigKey<Integer> TimeBetweenFailures = new ConfigKey<>("Advanced", Integer.class,
+    ConfigKey<Long> TimeBetweenFailures = new ConfigKey<>("Advanced", Long.class,
         "time.between.failures", valueOf(3600), "Time in seconds before try to cleanup all the VMs"
         + " which are registered for the HA event that were successful and are now ready to be purged.", false, null);
